requirements.txt
tuya-connector-python==0.13.0
python-dotenv==1.0.0